+++ src/main/java/laddergame/domain/ladder/Point.java
@@ -1,4 +1,4 @@
-package laddergame.domain;
+package laddergame.domain.ladder;
 
 import java.util.Objects;
