+++ src/main/java/laddergame/domain/ladder/RandomSelectStrategy.java
@@ -1,4 +1,4 @@
-package laddergame.domain;
+package laddergame.domain.ladder;
 
 import java.util.Random;
