+++ src/main/java/laddergame/domain/ladder/SelectStrategy.java
@@ -1,4 +1,4 @@
-package laddergame.domain;
+package laddergame.domain.ladder;
 
 public interface SelectStrategy {
     public boolean canSelect();
